requirements.txt
streamlit
pandas
numpy
scikit-learn
python-dotenv